Added a lastReply field for StepNode
