+++ stepik-java-api/src/main/java/org/stepik/api/objects/submissions/SubmissionPost.java
@@ -1,6 +1,7 @@
 package org.stepik.api.objects.submissions;
 
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * @author meanmail
@@ -15,6 +16,10 @@ public class SubmissionPost {
             reply = new Reply();
         }
         return reply;
+    }
+
+    public void setReply(@Nullable Reply reply) {
+        this.reply = reply;
     }
 
     public long getAttempt() {
